Product Type 변경 (Integer -> Double)

+++ process-service/src/main/java/com/poscoict/mes/process/product/jpa/ProductLogEntity.java
@@ -22,8 +22,8 @@ public class ProductLogEntity {
 	
 	private String productId;
 	private Double thickness;
-	private Integer width;
-	private Integer length;
+	private Double width;
+	private Double length;
 	private String productStatus;
 	private String productUpdate;
