+++ process-service/src/main/java/com/poscoict/mes/process/product/vo/ResponseProductLog.java
@@ -10,8 +10,8 @@ public class ResponseProductLog {
 	
 	private String productId;
 	private Double thickness;
-	private Integer width;
-	private Integer length;
+	private Double width;
+	private Double length;
 	private String productStatus;
 	private String productUpdate;
